spliting out LayerActionsDialog class into its own file
